% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catacode-package.R
\docType{package}
\name{catacode-package}
\alias{catacode}
\alias{catacode-package}
\title{catacode: Explore and code responses to check-all-that-apply survey items}
\description{
Analyzing responses to check-all-that-apply survey items often requires data transformations and subjective decisions for combining categories. 'catacode' contains tools for exploring response patterns, facilitating data transformations, applying a set of decision rules for coding responses, and summarizing response frequencies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/knickodem/catacode}
  \item Report bugs at \url{https://github.com/knickodem/catacode/issues}
}

}
\author{
\strong{Maintainer}: Kyle Nickodem \email{kyle.nickodem@gmail.com} (\href{https://orcid.org/0000-0003-4976-3378}{ORCID})

Authors:
\itemize{
  \item Gabriel J. Merrin \email{gjmerrin@syr.edu}
}

}
\keyword{internal}
