% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATAcode-package.R
\docType{package}
\name{CATAcode}
\alias{CATAcode}
\alias{CATAcode-package}
\title{CATAcode: Explore and Code Responses to Check-All-that-Apply Survey Items}
\description{
Analyzing responses to check-all-that-apply survey items often requires
data transformations and subjective decisions for combining categories. CATAcode
contains tools for exploring response patterns, facilitating data transformations,
applying a set of decision rules for coding responses, and summarizing response frequencies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/knickodem/CATAcode}
  \item \url{https://knickodem.github.io/CATAcode/}
  \item Report bugs at \url{https://github.com/knickodem/CATAcode/issues}
}

}
\author{
\strong{Maintainer}: Kyle Nickodem \email{kyle.nickodem@gmail.com} (\href{https://orcid.org/0000-0003-4976-3378}{ORCID})

Authors:
\itemize{
  \item Gabriel J. Merrin \email{gjmerrin@syr.edu}
}

}
