% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sources_race}
\alias{sources_race}
\title{Sources of Strength Race/Ethnicity Data}
\format{
A data frame with 16,922 rows and 9 columns:
\describe{
\item{ID}{Subject identification number}
\item{Wave}{Data collection time point}
\item{Black, Native_American, Asian, Hispanic, Multiracial, Pacific_Islander, White}{Indicator variables for check-all-that-apply
responses where 1 = endorsement}
}
}
\source{
\doi{doi:10.15139/S3/EZ8ILC}
}
\usage{
sources_race
}
\description{
Responses to the check-all-that-apply race/ethnicity question
at four time points from a randomized controlled trial of the
Sources of Strength program.
}
\keyword{datasets}
